+++ fuzzylite/types.py
@@ -1,3 +1,14 @@
+"""pyfuzzylite: a fuzzy logic control library in Python.
+
+This file is part of pyfuzzylite.
+
+Repository: https://github.com/fuzzylite/pyfuzzylite/
+
+License: FuzzyLite License
+
+Copyright: FuzzyLite by Juan Rada-Vilela. All rights reserved.
+"""
+
 from __future__ import annotations
 
 __all__ = ["Array", "Scalar", "ScalarArray"]