+++ tests/test_defuzzifier.py
@@ -1,20 +1,14 @@
-"""pyfuzzylite (TM), a fuzzy logic control library in Python.
-
-Copyright (C) 2010-2023 FuzzyLite Limited. All rights reserved.
-Author: Juan Rada-Vilela, PhD <jcrada@fuzzylite.com>.
+"""pyfuzzylite: a fuzzy logic control library in Python.
 
 This file is part of pyfuzzylite.
 
-pyfuzzylite is free software: you can redistribute it and/or modify it under
-the terms of the FuzzyLite License included with the software.
+Repository: https://github.com/fuzzylite/pyfuzzylite/
 
-You should have received a copy of the FuzzyLite License along with
-pyfuzzylite. If not, see <https://github.com/fuzzylite/pyfuzzylite/>.
+License: FuzzyLite License
 
-pyfuzzylite is a trademark of FuzzyLite Limited.
-
-fuzzylite is a registered trademark of FuzzyLite Limited.
+Copyright: FuzzyLite by Juan Rada-Vilela. All rights reserved.
 """
+
 from __future__ import annotations
 
 import re
@@ -89,12 +83,7 @@ class DefuzzifierAssert(BaseAssert[fl.Defuzzifier]):
         return self
 
 
-class NanTerm(fl.Term):
-    """A term that always returns NaN as its membership value."""
-
-    def membership(self, x: Scalar) -> Scalar:
-        """Returns NaN as the membership value of the term."""
-        return np.full_like(x, np.nan)
+NaN = fl.Constant
 
 
 class TestDefuzzifier(unittest.TestCase):
@@ -157,7 +146,7 @@ class TestDefuzzifier(unittest.TestCase):
         DefuzzifierAssert(self, fl.Bisector()).defuzzifies(
             0,
             2,
-            {NanTerm(): np.nan},  # mean of range of variable
+            {NaN(): np.nan},  # mean of range of variable
         )
 
     def test_centroid(self) -> None:
@@ -207,7 +196,7 @@ class TestDefuzzifier(unittest.TestCase):
         DefuzzifierAssert(self, fl.Centroid()).defuzzifies(
             -1,
             1,
-            {NanTerm(): np.nan},
+            {NaN(): np.nan},
         )
 
     def test_som_defuzzifier(self) -> None:
@@ -277,7 +266,7 @@ class TestDefuzzifier(unittest.TestCase):
         DefuzzifierAssert(self, fl.SmallestOfMaximum()).defuzzifies(
             -1,
             1,
-            {NanTerm(): np.nan},
+            {NaN(): np.nan},
         )
 
     def test_lom_defuzzifier(self) -> None:
@@ -349,7 +338,7 @@ class TestDefuzzifier(unittest.TestCase):
         DefuzzifierAssert(self, fl.LargestOfMaximum()).defuzzifies(
             -1,
             1,
-            {NanTerm(): np.nan},
+            {NaN(): np.nan},
         )
 
     def test_mom_defuzzifier(self) -> None:
@@ -420,7 +409,7 @@ class TestDefuzzifier(unittest.TestCase):
         DefuzzifierAssert(self, fl.MeanOfMaximum()).defuzzifies(
             -1,
             1,
-            {NanTerm(): np.nan},
+            {NaN(): np.nan},
         )
 
     def test_weighted_defuzzifier(self) -> None:
@@ -621,7 +610,7 @@ class TestDefuzzifier(unittest.TestCase):
                         fl.Activated(fl.Constant("C", -3.0), 0.5),
                     ],
                     aggregation=fl.AlgebraicSum(),
-                ): ((1 * 1 + 1 * -2 + (0.5 + 0.5 - (0.25)) * -3) / (1 + 1 + (0.5 + 0.5 - 0.25))),
+                ): -1.181818,  # ((1 * 1 + 1 * -2 + (0.5 + 0.5 - (0.25)) * -3) / (1 + 1 + (0.5 + 0.5 - 0.25))),
             },
             vectorized=False,
         )