+++ tests/test_library.py
@@ -1,20 +1,14 @@
-"""pyfuzzylite (TM), a fuzzy logic control library in Python.
-
-Copyright (C) 2010-2023 FuzzyLite Limited. All rights reserved.
-Author: Juan Rada-Vilela, PhD <jcrada@fuzzylite.com>.
+"""pyfuzzylite: a fuzzy logic control library in Python.
 
 This file is part of pyfuzzylite.
 
-pyfuzzylite is free software: you can redistribute it and/or modify it under
-the terms of the FuzzyLite License included with the software.
+Repository: https://github.com/fuzzylite/pyfuzzylite/
 
-You should have received a copy of the FuzzyLite License along with
-pyfuzzylite. If not, see <https://github.com/fuzzylite/pyfuzzylite/>.
+License: FuzzyLite License
 
-pyfuzzylite is a trademark of FuzzyLite Limited.
-
-fuzzylite is a registered trademark of FuzzyLite Limited.
+Copyright: FuzzyLite by Juan Rada-Vilela. All rights reserved.
 """
+
 from __future__ import annotations
 
 import logging
@@ -31,7 +25,7 @@ class TestLibrary(unittest.TestCase):
     def test_library_exports_dir(self) -> None:
         """Test the library exports expected components."""
         expected = """
-__builtins__ __cached__ __doc__ __file__ __loader__
+__author__ __builtins__ __cached__ __copyright__ __doc__ __file__ __license__ __loader__
 __name__ __package__ __path__ __spec__ __version__
 
 activation Activation First General Highest Last Lowest Proportional Threshold
@@ -76,10 +70,19 @@ variable InputVariable OutputVariable Variable
 
     def test_library_vars(self) -> None:
         """Test the library variables."""
-        __version__ = "8.0.2"
-        self.assertEqual(fl.__name__, "fuzzylite")
-        self.assertEqual(fl.__version__, __version__)
-        self.assertEqual(fl.__doc__, fl.information.description)
+        __version__ = "8.0.3"
+        self.assertTrue("fuzzylite" == fl.__name__ == fl.information.name)
+        self.assertTrue(__version__ == fl.__version__ == fl.information.version)
+        self.assertTrue(
+            "a fuzzy logic control library in Python" == fl.__doc__ == fl.information.description
+        )
+        self.assertTrue("Juan Rada-Vilela, PhD" == fl.__author__ == fl.information.author)
+        self.assertTrue("FuzzyLite License" == fl.__license__ == fl.information.license)
+        self.assertTrue(
+            "Copyright (C) 2010-2024 FuzzyLite by Juan Rada-Vilela. All rights reserved."
+            == fl.__copyright__
+            == fl.information.copyright
+        )
 
     def test_context(self) -> None:
         """Tests the context."""