+++ tests/test_norm.py
@@ -1,27 +1,23 @@
-"""pyfuzzylite (TM), a fuzzy logic control library in Python.
-
-Copyright (C) 2010-2023 FuzzyLite Limited. All rights reserved.
-Author: Juan Rada-Vilela, PhD <jcrada@fuzzylite.com>.
+"""pyfuzzylite: a fuzzy logic control library in Python.
 
 This file is part of pyfuzzylite.
 
-pyfuzzylite is free software: you can redistribute it and/or modify it under
-the terms of the FuzzyLite License included with the software.
+Repository: https://github.com/fuzzylite/pyfuzzylite/
 
-You should have received a copy of the FuzzyLite License along with
-pyfuzzylite. If not, see <https://github.com/fuzzylite/pyfuzzylite/>.
+License: FuzzyLite License
 
-pyfuzzylite is a trademark of FuzzyLite Limited.
-
-fuzzylite is a registered trademark of FuzzyLite Limited.
+Copyright: FuzzyLite by Juan Rada-Vilela. All rights reserved.
 """
+
 from __future__ import annotations
 
 import unittest
 
 import numpy as np
+import numpy.testing as npt
 
 import fuzzylite as fl
+from fuzzylite import inf, nan
 from tests.assert_component import BaseAssert
 
 
@@ -46,13 +42,20 @@ class NormAssert(BaseAssert[fl.Norm]):
     ) -> NormAssert:
         """Assert the norm produces the expected values."""
         for ab, z in abz.items():
-            self.test.assertEqual(z, self.actual.compute(*ab), f"in ({ab})")
+            npt.assert_allclose(z, self.actual.compute(*ab), equal_nan=True, err_msg=f"in ({ab})")
+            # self.test.assertEqual(z, self.actual.compute(*ab), f"in ({ab})")
             if commutative:
-                self.test.assertEqual(
+                npt.assert_allclose(
                     z,
                     self.actual.compute(*reversed(ab)),
-                    f"when ({tuple(reversed(ab))})",
+                    equal_nan=True,
+                    err_msg=f"when ({tuple(reversed(ab))})",
                 )
+                # self.test.assertEqual(
+                #     z,
+                #     self.actual.compute(*reversed(ab)),
+                #     f"when ({tuple(reversed(ab))})",
+                # )
             if associative:
                 a: fl.Scalar = ab[0]
                 b: fl.Scalar = ab[1]
@@ -103,6 +106,10 @@ class TestTNorm(unittest.TestCase):
                 (1.00, 0.50): 0.50,
                 (1.00, 0.75): 0.75,
                 (1.00, 1.00): 1.00,
+                (nan, nan): nan,
+                (inf, inf): inf,
+                (inf, -inf): -inf,
+                (-inf, -inf): inf,
             }
         )
 
@@ -125,6 +132,10 @@ class TestTNorm(unittest.TestCase):
                 (1.00, 0.50): 0.50,
                 (1.00, 0.75): 0.75,
                 (1.00, 1.00): 1.00,
+                (nan, nan): nan,
+                (inf, inf): inf,
+                (inf, -inf): nan,
+                (-inf, -inf): 0,
             }
         )
 
@@ -147,6 +158,10 @@ class TestTNorm(unittest.TestCase):
                 (1.00, 0.50): 0.50,
                 (1.00, 0.75): 0.75,
                 (1.00, 1.00): 1.00,
+                (nan, nan): 0.0,
+                (inf, inf): 0.0,
+                (inf, -inf): 0.0,
+                (-inf, -inf): 0.0,
             }
         )
 
@@ -169,6 +184,10 @@ class TestTNorm(unittest.TestCase):
                 (1.00, 0.50): 0.50,
                 (1.00, 0.75): 0.75,
                 (1.00, 1.00): 1.00,
+                (nan, nan): nan,
+                (inf, inf): nan,
+                (inf, -inf): nan,
+                (-inf, -inf): nan,
             }
         )
 
@@ -191,6 +210,10 @@ class TestTNorm(unittest.TestCase):
                 (1.00, 0.50): 0.50,
                 (1.00, 0.75): 0.75,
                 (1.00, 1.00): 1.00,
+                (nan, nan): nan,
+                (inf, inf): nan,
+                (inf, -inf): nan,
+                (-inf, -inf): nan,
             }
         )
 
@@ -213,6 +236,10 @@ class TestTNorm(unittest.TestCase):
                 (1.00, 0.50): 0.50,
                 (1.00, 0.75): 0.75,
                 (1.00, 1.00): 1.00,
+                (nan, nan): nan,
+                (inf, inf): inf,
+                (inf, -inf): -inf,
+                (-inf, -inf): -inf,
             }
         )
 
@@ -235,6 +262,10 @@ class TestTNorm(unittest.TestCase):
                 (1.00, 0.50): 0.50,
                 (1.00, 0.75): 0.75,
                 (1.00, 1.00): 1.00,
+                (nan, nan): 0.0,
+                (inf, inf): inf,
+                (inf, -inf): 0.0,
+                (-inf, -inf): 0.0,
             }
         )
 
@@ -261,6 +292,10 @@ class TestSNorm(unittest.TestCase):
                 (1.00, 0.50): 1.00,
                 (1.00, 0.75): 1.00,
                 (1.00, 1.00): 1.00,
+                (nan, nan): nan,
+                (inf, inf): nan,
+                (inf, -inf): nan,
+                (-inf, -inf): -inf,
             }
         )
 
@@ -283,6 +318,10 @@ class TestSNorm(unittest.TestCase):
                 (1.00, 0.50): 1.00,
                 (1.00, 0.75): 1.00,
                 (1.00, 1.00): 1.00,
+                (nan, nan): nan,
+                (inf, inf): 1.0,
+                (inf, -inf): nan,
+                (-inf, -inf): -inf,
             }
         )
 
@@ -305,6 +344,10 @@ class TestSNorm(unittest.TestCase):
                 (1.00, 0.50): 1.00,
                 (1.00, 0.75): 1.00,
                 (1.00, 1.00): 1.00,
+                (nan, nan): 1.0,
+                (inf, inf): 1.0,
+                (inf, -inf): 1.0,
+                (-inf, -inf): 1.0,
             }
         )
 
@@ -327,6 +370,10 @@ class TestSNorm(unittest.TestCase):
                 (1.00, 0.50): 1.00,
                 (1.00, 0.75): 1.00,
                 (1.00, 1.00): 1.00,
+                (nan, nan): nan,
+                (inf, inf): nan,
+                (inf, -inf): nan,
+                (-inf, -inf): nan,
             }
         )
 
@@ -349,6 +396,10 @@ class TestSNorm(unittest.TestCase):
                 (1.00, 0.50): 1.00,
                 (1.00, 0.75): 1.00,
                 (1.00, 1.00): 1.00,
+                (nan, nan): nan,
+                (inf, inf): nan,
+                (inf, -inf): nan,
+                (-inf, -inf): nan,
             }
         )
 
@@ -371,6 +422,10 @@ class TestSNorm(unittest.TestCase):
                 (1.00, 0.50): 1.00,
                 (1.00, 0.75): 1.00,
                 (1.00, 1.00): 1.00,
+                (nan, nan): nan,
+                (inf, inf): inf,
+                (inf, -inf): inf,
+                (-inf, -inf): -inf,
             }
         )
 
@@ -393,6 +448,10 @@ class TestSNorm(unittest.TestCase):
                 (1.00, 0.50): 1.00,
                 (1.00, 0.75): 1.00,
                 (1.00, 1.00): 1.00,
+                (nan, nan): 1.0,
+                (inf, inf): 1.0,
+                (inf, -inf): 1.0,
+                (-inf, -inf): -inf,
             }
         )
 
@@ -415,6 +474,10 @@ class TestSNorm(unittest.TestCase):
                 (1.00, 0.50): 1.00,
                 (1.00, 0.75): 1.00,
                 (1.00, 1.00): 1.00,
+                (nan, nan): nan,
+                (inf, inf): nan,
+                (inf, -inf): nan,
+                (-inf, -inf): -inf,
             }
         )
 
@@ -437,6 +500,10 @@ class TestSNorm(unittest.TestCase):
                 (1.00, 0.50): 1.50,
                 (1.00, 0.75): 1.75,
                 (1.00, 1.00): 2.00,
+                (nan, nan): nan,
+                (inf, inf): inf,
+                (inf, -inf): nan,
+                (-inf, -inf): -inf,
             }
         )
 
@@ -465,6 +532,10 @@ class TestNormFunctions(unittest.TestCase):
                 (1.00, 0.50): 1.00,
                 (1.00, 0.75): 1.00,
                 (1.00, 1.00): 1.00,
+                (nan, nan): nan,
+                (inf, inf): nan,
+                (inf, -inf): nan,
+                (-inf, -inf): -inf,
             }
         )
 
@@ -487,6 +558,11 @@ class TestNormFunctions(unittest.TestCase):
                 (1.00, 0.50): 1.00,
                 (1.00, 0.75): 1.00,
                 (1.00, 1.00): 1.00,
+                (nan, nan): nan,
+                (inf, inf): nan,
+                (-inf, inf): nan,
+                (inf, -inf): nan,
+                (-inf, -inf): -inf,
             }
         )