+++ tests/test_rule.py
@@ -1,20 +1,14 @@
-"""pyfuzzylite (TM), a fuzzy logic control library in Python.
-
-Copyright (C) 2010-2023 FuzzyLite Limited. All rights reserved.
-Author: Juan Rada-Vilela, PhD <jcrada@fuzzylite.com>.
+"""pyfuzzylite: a fuzzy logic control library in Python.
 
 This file is part of pyfuzzylite.
 
-pyfuzzylite is free software: you can redistribute it and/or modify it under
-the terms of the FuzzyLite License included with the software.
+Repository: https://github.com/fuzzylite/pyfuzzylite/
 
-You should have received a copy of the FuzzyLite License along with
-pyfuzzylite. If not, see <https://github.com/fuzzylite/pyfuzzylite/>.
+License: FuzzyLite License
 
-pyfuzzylite is a trademark of FuzzyLite Limited.
-
-fuzzylite is a registered trademark of FuzzyLite Limited.
+Copyright: FuzzyLite by Juan Rada-Vilela. All rights reserved.
 """
+
 from __future__ import annotations
 
 import unittest
@@ -127,8 +121,9 @@ class AssertAntecedent:
 
     def has_activation_degrees(
         self,
-        inputs: dict[fl.InputVariable, list[float]]
-        | dict[fl.OutputVariable, list[list[fl.Activated]]],
+        inputs: (
+            dict[fl.InputVariable, list[float]] | dict[fl.OutputVariable, list[list[fl.Activated]]]
+        ),
         rules: dict[str, list[float]],
         conjunction: fl.TNorm | None = None,
         disjunction: fl.SNorm | None = None,
@@ -993,6 +988,10 @@ class TestRuleBlock(unittest.TestCase):
         """Test iter, len, and getitem of rule blocks."""
         a, b = [fl.Rule.create("if a then z"), fl.Rule.create("if b then y")]
         rule_block = fl.RuleBlock("test", rules=[a, b])
+
+        # test rule()
+        self.assertEqual(rule_block.rule(0), a)
+        self.assertEqual(rule_block.rule(1), b)
 
         # test getitem
         self.assertEqual(rule_block[0], a)